feat: split moderator commands into private cli
