+++ packages/clawhub-mod/bin/clawhub-mod.js
@@ -1,0 +1,2 @@
+#!/usr/bin/env node
+import "../dist/clawhub-mod/src/cli.js";
