+++ packages/clawhub-mod/package.json
@@ -1,0 +1,49 @@
+{
+  "name": "@openclaw/clawhub-mod",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Moderator-only ClawHub operator CLI for platform moderation and package operations.",
+  "homepage": "https://clawhub.ai",
+  "license": "UNLICENSED",
+  "repository": {
+    "type": "git",
+    "url": "https://github.com/openclaw/clawhub.git",
+    "directory": "packages/clawhub-mod"
+  },
+  "bin": {
+    "clawhub-mod": "bin/clawhub-mod.js"
+  },
+  "files": [
+    "bin",
+    "dist",
+    "README.md"
+  ],
+  "type": "module",
+  "scripts": {
+    "build": "node ./scripts/build.mjs",
+    "dev": "node --enable-source-maps dist/clawhub-mod/src/cli.js",
+    "test": "vitest run -c vitest.config.ts",
+    "typecheck": "node ./scripts/typecheck.mjs",
+    "verify": "bun run test && bun run typecheck && bun run build"
+  },
+  "dependencies": {
+    "@clack/prompts": "^1.3.0",
+    "arktype": "^2.2.0",
+    "commander": "^14.0.3",
+    "fflate": "^0.8.2",
+    "ignore": "^7.0.5",
+    "json5": "^2.2.3",
+    "mime": "^4.1.0",
+    "ora": "^9.4.0",
+    "p-retry": "8.0.0",
+    "semver": "^7.7.4",
+    "undici": "7.25.0"
+  },
+  "devDependencies": {
+    "@types/node": "^25.5.0",
+    "typescript": "6.0.3"
+  },
+  "engines": {
+    "node": ">=20"
+  }
+}
