+++ packages/clawhub-mod/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "compilerOptions": {
+    "target": "ES2022",
+    "module": "ES2022",
+    "moduleResolution": "Bundler",
+    "outDir": "dist",
+    "rootDir": "..",
+    "strict": true,
+    "declaration": true,
+    "sourceMap": true,
+    "skipLibCheck": true
+  },
+  "include": ["src/**/*.ts", "../clawhub/src/**/*.ts"],
+  "exclude": ["src/**/*.test.ts", "../clawhub/src/**/*.test.ts"]
+}
