+++ tsconfig.oxlint.json
@@ -1,5 +1,11 @@
 {
-  "include": ["src", "convex", "packages/clawhub/src", "packages/schema/src"],
+  "include": [
+    "src",
+    "convex",
+    "packages/clawhub/src",
+    "packages/clawhub-mod/src",
+    "packages/schema/src"
+  ],
   "compilerOptions": {
     "target": "ES2022",
     "jsx": "react-jsx",
